Move back paginator class to framework package
