+++ packages/framework/resources/views/components/publications/pagination-navigation.blade.php
@@ -1,4 +1,4 @@
-@php/** @var \Hyde\Publications\Paginator $paginator */@endphp
+@php/** @var \Hyde\Framework\Features\Paginator $paginator */@endphp
 <nav class="flex justify-between mt-4">
     @if($paginator->previous())
         <x-link :href="$paginator->previous()">Prev</x-link>
