+++ packages/publications/tests/PublicationTypeTest.php
@@ -11,7 +11,7 @@ use Hyde\Publications\Models\PublicationFieldDefinition;
 use Hyde\Publications\Models\PublicationListPage;
 use Hyde\Publications\Models\PublicationPage;
 use Hyde\Publications\Models\PublicationType;
-use Hyde\Publications\Paginator;
+use Hyde\Framework\Features\Paginator;
 use Hyde\Publications\PublicationService;
 use Hyde\Testing\TestCase;
 use Illuminate\Support\Collection;
